auth-service/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
# PostgreSQL driver - needed for AWS RDS
psycopg2-binary==2.9.10
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
# For health check in Docker
requests==2.31.0
